:truck: move code to mod tools
